Levels (#114)
* Adding level up logic (WIP)

* Extras
