+++ client/src/game/systems/camera/CameraMovementSystem.java
@@ -4,7 +4,9 @@ import camera.AOCamera;
 import com.artemis.Aspect;
 import com.artemis.annotations.Wire;
 import com.artemis.systems.IteratingSystem;
+import com.badlogic.gdx.math.MathUtils;
 import position.Pos2D;
+import shared.model.map.Tile;
 
 import static com.artemis.E.E;
 
@@ -24,8 +26,9 @@ public class CameraMovementSystem extends IteratingSystem {
     protected void process(int camera) {
         final Pos2D pos = E(camera).getPos2D();
 
-        cameraSystem.camera.position.x = (pos.x);
-        cameraSystem.camera.position.y = (pos.y);
+        cameraSystem.camera.position.x = pos.x;
+        cameraSystem.camera.position.y = pos.y;
+        cameraSystem.camera.position.x += Tile.TILE_PIXEL_WIDTH / 2;
         cameraSystem.camera.update();
     }
