+++ components/src/entity/character/info/Gold.java
@@ -1,0 +1,25 @@
+package entity.character.info;
+
+import com.artemis.Component;
+import com.artemis.annotations.PooledWeaver;
+
+import java.io.Serializable;
+
+@PooledWeaver
+public class Gold extends Component implements Serializable {
+    int count;
+
+    public Gold() {}
+
+    public Gold(int count) {
+        this.count = count;
+    }
+
+    public void setCount(int count) {
+        this.count = count;
+    }
+
+    public int getCount() {
+        return count;
+    }
+}
