+++ server/resources/hechizos.dat
