+++ server/resources/obj.dat
