+++ server/src/server/network/ServerRequestProcessor.java
@@ -9,8 +9,8 @@ import entity.world.Dialog;
 import entity.world.Object;
 import movement.Destination;
 import position.WorldPos;
-import server.combat.MagicCombatSystem;
-import server.combat.PhysicalCombatSystem;
+import server.systems.combat.MagicCombatSystem;
+import server.systems.combat.PhysicalCombatSystem;
 import server.systems.MeditateSystem;
 import server.systems.ServerSystem;
 import server.systems.manager.ItemManager;
