+++ server/src/server/systems/ai/NPCAttackSystem.java
@@ -6,8 +6,8 @@ import entity.character.states.Heading;
 import entity.character.status.Hit;
 import entity.npc.NPC;
 import position.WorldPos;
-import server.combat.CombatSystem;
-import server.combat.PhysicalCombatSystem;
+import server.systems.combat.CombatSystem;
+import server.systems.combat.PhysicalCombatSystem;
 import server.systems.IntervalFluidIteratingSystem;
 import server.systems.manager.MapManager;
 import server.systems.manager.WorldManager;
