+++ server/src/server/systems/manager/WorldManager.java
@@ -4,31 +4,17 @@ import camera.Focused;
 import com.artemis.Component;
 import com.artemis.E;
 import com.artemis.annotations.Wire;
-import com.esotericsoftware.minlog.Log;
 import entity.character.states.CanWrite;
-import entity.character.states.Heading;
-import entity.character.status.Hit;
 import entity.npc.OriginPos;
 import physics.AOPhysics;
-import position.WorldPos;
-import server.core.Server;
-import server.database.model.attributes.Attributes;
-import server.database.model.modifiers.Modifiers;
+import server.systems.EntityFactorySystem;
 import server.systems.ServerSystem;
-import shared.interfaces.CharClass;
-import shared.interfaces.Hero;
-import shared.interfaces.Race;
-import shared.model.Spell;
 import shared.model.lobby.Player;
-import shared.model.lobby.Team;
 import shared.model.npcs.NPC;
 import shared.network.notifications.EntityUpdate.EntityUpdateBuilder;
 import shared.network.notifications.RemoveEntity;
-import shared.objects.types.*;
 
-import java.util.*;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.stream.Collectors;
+import java.util.List;
 
 import static com.artemis.E.E;
 import static server.utils.WorldUtils.WorldUtils;
@@ -40,565 +26,7 @@ public class WorldManager extends DefaultManager {
     private ServerSystem networkManager;
     private SpellManager spellManager;
     private ObjectManager objectManager;
-
-    private static final int ATTR_BASE_VALUE = 18;
-    private static int MAX_LEVEL = 45;
-    private static int STAT_MAXHIT_UNDER36 = 99;
-    private static int STAT_MAXHIT_OVER36 = 999;
-
-    public void createObject(int objIndex, int objCount, WorldPos pos) {
-        int objId = world.create();
-        E(objId).worldPosMap(pos.map)
-                .worldPosX(pos.x)
-                .worldPosY(pos.y)
-                .objectIndex(objIndex)
-                .objectCount(objCount);
-        registerEntity(objId);
-    }
-
-    public void createNPC(int npcIndex, WorldPos pos) {
-        NPC npc = world.getSystem(NPCManager.class).getNpcs().get(npcIndex);
-        int npcId = world.create();
-
-        E npcEntity = E(npcId);
-        npcEntity
-                .nPCId(npcIndex)
-                .bodyIndex(npc.getBody())
-                .headingCurrent(Heading.HEADING_SOUTH)
-                .nameText(npc.getName());
-        if (npc.getMovement() == 3) {
-            npcEntity.aOPhysics().aOPhysicsVelocity(85f);
-            npcEntity.aIMovement();
-        }
-        if (npc.getHead() > 0) {
-            npcEntity.headIndex(npc.getHead());
-        }
-        if (npc.isCommerce()) {
-            npcEntity.commerce();
-        }
-        if (npc.isHostile()) {
-            npcEntity.hostile();
-        }
-        if (npc.getMaxHit() > 0) {
-            npcEntity.hit().hitMax(npc.getMaxHit()).hitMin(npc.getMinHit());
-        }
-        if (npc.getEvasionPower() > 0) {
-            npcEntity.evasionPowerValue(npc.getEvasionPower());
-        }
-        if (npc.getAttackPower() > 0) {
-            npcEntity.attackPowerValue(npc.getAttackPower());
-        }
-        if (npc.getMaxHP() > 0) {
-            npcEntity.health().healthMin(npc.getMinHP()).healthMax(npc.getMaxHP());
-        }
-        if (npc.isAttackable()) {
-            npcEntity.attackable();
-        }
-
-        npcEntity.originPosMap(pos.map).originPosX(pos.x).originPosY(pos.y);
-        npcEntity.worldPosMap(pos.map).worldPosX(pos.x).worldPosY(pos.y);
-        registerEntity(npcId);
-    }
-
-    public int createEntity(String name, Hero hero, Team team) {
-        int player = getWorld().create();
-
-        E entity = E(player);
-        entity.character();
-        switch (team) {
-            case NO_TEAM:
-                entity.gM();
-                break;
-            case CAOS_ARMY:
-                entity.criminal();
-                break;
-        }
-        // set position
-        setEntityPosition(entity);
-        // set head and body
-        setHeadAndBody(name, entity);
-        // set class
-        setClassAndAttributes(hero, entity);
-        // set inventory
-        setInventory(player, hero, team);
-        // set spells
-        setSpells(player, hero);
-
-        return player;
-    }
-
-    private void setSpells(int player, Hero hero) {
-        Set<Spell> spells = getSpells(hero);
-        final Integer[] spellIds = spells
-                .stream()
-                .map(spell -> spellManager.getId(spell))
-                .toArray(Integer[]::new);
-        E(player).spellBookSpells(spellIds);
-    }
-
-    private Set<Spell> getSpells(Hero hero) {
-        final Map<Integer, Spell> spells = spellManager.getSpells();
-        Set<Spell> result = new HashSet<>();
-        Spell apoca = spells.get(25);
-        Spell desca = spells.get(23);
-        Spell tormenta = spells.get(15);
-        Spell misil = spells.get(8);
-        Spell inmo = spells.get(24);
-        Spell remo = spells.get(10);
-        switch (hero) {
-            case MAGO:
-            case BARDO:
-            case CLERIGO:
-                result.add(apoca);
-                break;
-        }
-        if (!(hero.equals(Hero.GUERRERO) || hero.equals(Hero.ARQUERO))) {
-            result.add(misil);
-            result.add(tormenta);
-            result.add(desca);
-            result.add(inmo);
-            result.add(remo);
-        }
-        return result;
-    }
-
-    private void setClassAndAttributes(Hero hero, E entity) {
-        // set body and head
-
-        Race race = Race.values()[hero.getRaceId()];
-        setNakedBody(entity, race);
-        setHead(entity, race);
-        entity.charHeroHeroId(hero.ordinal());
-
-        CharClass charClass = CharClass.values()[hero.getClassId()];
-        setAttributesAndStats(entity, charClass, race);
-    }
-
-    private void setAttributesAndStats(E entity, CharClass charClass, Race race) {
-        // set attributes
-        //        entity.agilityValue(ATTR_BASE_VALUE + Attributes.AGILITY.of(race));
-        entity.agilityBaseValue(38);
-        entity.agilityCurrentValue(38);
-        entity.charismaBaseValue(ATTR_BASE_VALUE + Attributes.CHARISMA.of(race));
-        entity.constitutionBaseValue(ATTR_BASE_VALUE + Attributes.CONSTITUTION.of(race));
-        entity.intelligenceBaseValue(ATTR_BASE_VALUE + Attributes.INTELLIGENCE.of(race));
-        //        entity.strengthValue(ATTR_BASE_VALUE + Attributes.STRENGTH.of(race));
-        entity.strengthBaseValue(38);
-        entity.strengthCurrentValue(38);
-
-        // set stats
-        setLevel(entity);
-        setStamina(entity);
-        setHP(entity, charClass);
-        setMana(entity, charClass);
-        setHit(entity, charClass);
-    }
-
-    private void setLevel(E entity) {
-        entity.levelLevel(MAX_LEVEL);
-    }
-
-    private void setStamina(E entity) {
-        // set stamina
-        entity.staminaMax(500);
-        entity.staminaMin(500);
-    }
-
-    private void setHit(E entity, CharClass charClass) {
-        int minLvl, maxLvl;
-        int breakingLvl = 35;
-        switch (charClass) {
-            case WARRIOR:
-            case ARCHER:
-                minLvl = 3;
-                maxLvl = 2;
-                break;
-            case PIRATE:
-                minLvl = 3;
-                maxLvl = 3;
-                break;
-            case PALADIN:
-            case ASSASSIN:
-            case ROGUE:
-                minLvl = 3;
-                maxLvl = 1;
-                break;
-            case CLERIC:
-            case BARDIC:
-            case DRUID:
-            case THIEF:
-                minLvl = 2;
-                maxLvl = 2;
-                break;
-            case MAGICIAN:
-                minLvl = 1;
-                maxLvl = 1;
-                break;
-            default:
-                minLvl = 0;
-                maxLvl = 0;
-                break;
-        }
-        int maxHit = Math.min(STAT_MAXHIT_OVER36,
-                Math.min(minLvl * breakingLvl + (entity.getLevel().level - breakingLvl) * maxLvl,
-                        STAT_MAXHIT_UNDER36));
-        entity.hit();
-        Hit hit = entity.getHit();
-        hit.setMax(maxHit);
-        hit.setMin(maxHit);
-    }
-
-    private void setHead(E entity, Race race) {
-        ThreadLocalRandom random = ThreadLocalRandom.current();
-        int headIndex = 0;
-        switch (race) {
-            case HUMAN:
-                headIndex = random.nextInt(1, 51 + 1);
-                break;
-            case DROW:
-                headIndex = random.nextInt(201, 221 + 1);
-                break;
-            case ELF:
-                headIndex = random.nextInt(101, 122 + 1);
-                break;
-            case GNOME:
-                headIndex = random.nextInt(401, 416 + 1);
-                break;
-            case DWARF:
-                headIndex = random.nextInt(301, 319 + 1);
-        }
-        entity.headIndex(headIndex);
-    }
-
-    void setNakedBody(E entity, Race race) {
-        int bodyIndex = 1;
-        switch (race) {
-            case GNOME:
-                bodyIndex = 222;
-                break;
-            case DWARF:
-                bodyIndex = 53;
-                break;
-            case ELF:
-                bodyIndex = 210;
-                break;
-            case DROW:
-                bodyIndex = 32;
-                break;
-            case HUMAN:
-                bodyIndex = 21;
-        }
-        entity.bodyIndex(bodyIndex);
-    }
-
-    private void setMana(E entity, CharClass heroClass) {
-        int manaBase = 50;
-        float manaPerLvlFactor;
-        switch (heroClass) {
-            case ROGUE:
-                manaPerLvlFactor = (float) (1 / 3) * 2;
-                break;
-            case PALADIN:
-            case ASSASSIN:
-                manaPerLvlFactor = 1;
-                break;
-            case MAGICIAN:
-                manaPerLvlFactor = 2.8f;
-                break;
-            case BARDIC:
-            case CLERIC:
-            case DRUID:
-                manaPerLvlFactor = 2f;
-                break;
-            default:
-                manaPerLvlFactor = 0;
-                break;
-        }
-        float manaPerLvl = entity.intelligenceBaseValue() * manaPerLvlFactor;
-        int maxMana = (int) (manaPerLvl * (entity.levelLevel() - 1) + manaBase);
-        maxMana = maxMana == manaBase ? 0 : maxMana;
-        entity.manaMax(maxMana);
-        entity.manaMin(maxMana);
-    }
-
-    private void setHP(E entity, CharClass charClass) {
-        int initialBase = 20 + ThreadLocalRandom.current().nextInt(-2, 2);
-        int average = (int) Modifiers.HEALTH.of(charClass) * entity.getLevel().level - 1;
-        int maxHP = initialBase + ThreadLocalRandom.current().nextInt(average - 15, average + 15);
-        entity.healthMax(maxHP);
-        entity.healthMin(maxHP);
-    }
-
-    private void setInventory(int player, Hero hero, Team team) {
-        E(player).inventory();
-        addPotion(player, PotionKind.HP);
-
-        if (E(player).manaMax() > 0) {
-            addPotion(player, PotionKind.MANA);
-        } else {
-            addPotion(player, PotionKind.STRENGTH);
-        }
-
-        getHelmet(hero, team).ifPresent(helmet -> {
-            E(player).helmetIndex(helmet.getId());
-            E(player).getInventory().add(helmet.getId(), true);
-        });
-        getArmor(hero, team).ifPresent(armor -> {
-            E(player).armorIndex(armor.getId());
-            E(player).bodyIndex(((ArmorObj) armor).getBodyNumber());
-            E(player).getInventory().add(armor.getId(), true);
-        });
-        final Set<Obj> weapons = getWeapon(hero, team);
-        if (!weapons.isEmpty()) {
-            final Obj next = weapons.iterator().next();
-            E(player).getInventory().add(next.getId(), true);
-            E(player).weaponIndex(next.getId());
-            weapons.forEach(weapon -> {
-                if (weapon != next) {
-                    E(player).getInventory().add(weapon.getId(), false);
-                }
-            });
-        }
-
-        getShield(hero, team).ifPresent(shield -> {
-            E(player).shieldIndex(shield.getId());
-            E(player).getInventory().add(shield.getId(), true);
-        });
-    }
-
-    private void setHeadAndBody(String name, E entity) {
-        entity
-                .headingCurrent(Heading.HEADING_NORTH)
-                .character()
-                .nameText(name);
-    }
-
-    private void setEntityPosition(E entity) {
-        WorldPos worldPos = getValidPosition(1);
-        entity
-                .worldPosX(worldPos.x)
-                .worldPosY(worldPos.y)
-                .worldPosMap(worldPos.map);
-    }
-
-    private WorldPos getValidPosition(int map) {
-        return new WorldPos(50, 50, map);
-    }
-
-    private WorldPos getRandomPos(int maxWidth, int maxHeight, int map) {
-        final int x = ThreadLocalRandom.current().nextInt(maxWidth);
-        final int y = ThreadLocalRandom.current().nextInt(maxHeight);
-        return new WorldPos(x, y, map);
-    }
-
-    private void addPotion(int player, PotionKind kind) {
-        Set<Obj> objs = objectManager.getTypeObjects(Type.POTION);
-        objs.stream() //
-                .map(PotionObj.class::cast) //
-                .filter(potion -> {
-                    PotionKind potionKind = potion.getKind();
-                    return potionKind != null && potionKind.equals(kind);
-                }) //
-                .findFirst() //
-                .ifPresent(obj -> E(player).getInventory().add(obj.getId(), false));
-    }
-
-    private Optional<Obj> getArmor(Hero hero, Team team) {
-        final Random random = new Random();
-        Optional<Obj> result = Optional.empty();
-        List<Integer> noTeam;
-        List<Integer> real;
-        List<Integer> chaos;
-        List<Integer> set;
-        switch (hero) {
-            case PALADIN:
-                noTeam = Arrays.asList(195, 485, 496);
-                real = Collections.singletonList(680);
-                chaos = Collections.singletonList(683);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-            case GUERRERO:
-                noTeam = Arrays.asList(243, 968);
-                real = Arrays.asList(681, 694);
-                chaos = Arrays.asList(685, 695);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-            case MAGO:
-                noTeam = Arrays.asList(525, 969);
-                real = Arrays.asList(549, 682);
-                chaos = Arrays.asList(558, 686);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-            case BARDO:
-                noTeam = Arrays.asList(519, 359, 484);
-                real = Collections.singletonList(520);
-                chaos = Collections.singletonList(523);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-            case ARQUERO:
-                noTeam = Arrays.asList(964, 965);
-                real = Collections.singletonList(1040);
-                chaos = Collections.singletonList(1041);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-            case ASESINO:
-                noTeam = Arrays.asList(356, 495);
-                real = Arrays.asList(521, 691);
-                chaos = Arrays.asList(684, 701);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-            case CLERIGO:
-                noTeam = Arrays.asList(356, 495);
-                real = Collections.singletonList(521);
-                chaos = Collections.singletonList(523);
-
-                set = team.equals(Team.NO_TEAM) ? noTeam : team.equals(Team.CAOS_ARMY) ? chaos : real;
-                result = objectManager.getObject(set.get(random.nextInt(set.size())));
-                break;
-        }
-        return result;
-    }
-
-    private Optional<Obj> getHelmet(Hero hero, Team team) {
-        Optional<Obj> result = Optional.empty();
-        switch (hero) {
-            case PALADIN:
-            case GUERRERO:
-                result = objectManager.getObject(405);
-                break;
-            case ARQUERO:
-                Random random = new Random();
-                result = objectManager.getObject(Arrays.asList(1052, 1003).get(random.nextInt(2)));
-                break;
-            case CLERIGO:
-            case ASESINO:
-                result = objectManager.getObject(131);
-                break;
-            case BARDO:
-            case MAGO:
-                result = objectManager.getObject(851);
-                break;
-            default:
-                break;
-        }
-        return result;
-    }
-
-    private Optional<Obj> getShield(Hero hero, Team team) {
-        Optional<Obj> result = Optional.empty();
-        switch (hero) {
-            case PALADIN:
-            case CLERIGO:
-            case GUERRERO:
-                result = objectManager.getObject(team.equals(Team.NO_TEAM) ? 130 : team.equals(Team.REAL_ARMY) ? 1038 : 1037);
-                break;
-            case BARDO:
-            case ASESINO:
-                result = objectManager.getObject(404);
-                break;
-            default:
-                break;
-        }
-        return result;
-    }
-
-    private Set<Obj> getWeapon(Hero hero, Team type) {
-        Set<Obj> result = new HashSet<>();
-        switch (hero) {
-            case PALADIN:
-            case GUERRERO:
-                result.add(objectManager.getObject(403).get());
-                break;
-            case ARQUERO:
-                result.add(objectManager.getObject(665).get());
-                result.add(objectManager.getObject(366).get());
-                break;
-            case BARDO:
-                result.add(objectManager.getObject(366).get());
-                break;
-            case CLERIGO:
-                result.add(objectManager.getObject(129).get());
-                break;
-            case ASESINO:
-                result.add(objectManager.getObject(559).get());
-                break;
-            case MAGO:
-                result.add(objectManager.getObject(660).get());
-                break;
-            default:
-                break;
-        }
-        return result;
-    }
-
-
-    private Optional<Obj> addItem(int player, Type type) {
-        Set<Obj> objs = objectManager.getTypeObjects(type);
-        Optional<Obj> result = objs.stream()
-                .filter(obj -> {
-                    if (obj instanceof ObjWithClasses) {
-                        int heroId = E(player).getCharHero().heroId;
-                        CharClass clazz = CharClass.get(E(player));
-                        Set<CharClass> forbiddenClasses = ((ObjWithClasses) obj).getForbiddenClasses();
-                        boolean supported = forbiddenClasses.size() == 0 || !forbiddenClasses.contains(clazz);
-                        if (supported && obj instanceof ArmorObj) {
-                            Race race = Race.of(E(player));
-                            if (race.equals(Race.GNOME) || race.equals(Race.DWARF)) {
-                                supported = ((ArmorObj) obj).isDwarf();
-                            } else if (((ArmorObj) obj).isWomen()) {
-                                supported = false; // TODO
-                            }
-                        }
-                        return supported;
-                    } else if (obj.getType().equals(Type.POTION)) {
-                        PotionObj potion = (PotionObj) obj;
-                        return potion.getKind().equals(PotionKind.HP) || potion.getKind().equals(PotionKind.MANA);
-                    }
-                    return false;
-                })
-                .max(this::getComparator);
-        result.ifPresent(obj -> {
-            CharClass clazz = CharClass.get(E(player));
-            Set<CharClass> forbiddenClasses = ((ObjWithClasses) obj).getForbiddenClasses();
-            Log.info("Item found for class: " + clazz.name() + " and forbidden classes are: " + Arrays
-                    .toString(forbiddenClasses.toArray()));
-            E(player).getInventory().add(obj.getId(), true);
-        });
-        return result;
-    }
-
-    private int getComparator(Obj obj1, Obj obj2) {
-        if (obj1 instanceof ArmorObj) {
-            ArmorObj armor1 = (ArmorObj) obj1;
-            ArmorObj armor2 = (ArmorObj) obj2;
-            return armor1.getMaxDef() - armor2.getMaxDef();
-        } else if (obj1 instanceof WeaponObj) {
-            WeaponObj weapon1 = (WeaponObj) obj1;
-            WeaponObj weapon2 = (WeaponObj) obj2;
-            return weapon1.getMaxHit() - weapon2.getMaxHit();
-        } else if (obj1 instanceof ShieldObj) {
-            ShieldObj shield1 = (ShieldObj) obj1;
-            ShieldObj shield2 = (ShieldObj) obj2;
-            return shield1.getMaxDef() - shield2.getMaxDef();
-        } else if (obj1 instanceof HelmetObj) {
-            HelmetObj helmet1 = (HelmetObj) obj1;
-            HelmetObj helmet2 = (HelmetObj) obj2;
-            return helmet1.getMaxDef() - helmet2.getMaxDef();
-        }
-        return obj1.getName().compareTo(obj2.getName());
-    }
+    private EntityFactorySystem entityFactorySystem;
 
     public void registerEntity(int id) {
         mapManager.updateEntity(id);
@@ -606,10 +34,11 @@ public class WorldManager extends DefaultManager {
 
     public void registerEntity(int connectionId, int id) {
         networkManager.registerUserConnection(id, connectionId);
-        mapManager.updateEntity(id);
+        registerEntity(id);
     }
 
     public void unregisterEntity(int entityId) {
+        networkManager.unregisterUserConnection(entityId);
         mapManager.removeEntity(entityId);
         getWorld().delete(entityId);
     }
@@ -639,7 +68,7 @@ public class WorldManager extends DefaultManager {
     }
 
     public void entityDie(int entityId) {
-        // RESET USER. TODO implement ghost
+
         final E e = E(entityId);
         if (e.hasNPC()) {
             int npcId = e.nPCId();
@@ -655,7 +84,7 @@ public class WorldManager extends DefaultManager {
 
             unregisterEntity(e.id());
         } else {
-            setEntityPosition(e);
+            // RESET USER. TODO implement ghost
             // reset health
             e.getHealth().min = e.getHealth().max;
             // reset mana
@@ -671,7 +100,7 @@ public class WorldManager extends DefaultManager {
     }
 
     public void login(int connectionId, Player player) {
-        final int entity = createEntity(player.getPlayerName(), player.getHero(), player.getTeam());
+        final int entity = entityFactorySystem.createPlayer(player.getPlayerName(), player.getHero(), player.getTeam());
         List<Component> components = WorldUtils(getWorld()).getComponents(getWorld().getEntity(entity));
         components.add(new Focused());
         components.add(new AOPhysics());
