+++ shared/build.gradle
@@ -18,8 +18,10 @@ ext {
     fluidOutputDir = file("$buildDir/generated-sources/fluid/")
 }
 
-sourceSets.main.java.srcDirs = ["src/", fluidOutputDir]
-sourceSets.main.resources.srcDirs = ["resources/"]
+sourceSets {
+	main.java.srcDirs = ["src/", fluidOutputDir]
+	main.resources.srcDirs = ["resources/"]
+}
 
 fluid {
     generatedSourcesDirectory = fluidOutputDir
@@ -37,15 +39,14 @@ fluid {
 }
 
 weave {
-    classesDir = new File(buildDir, "classes/java/main")
-    enableArtemisPlugin = true
+	classesDir = new File(buildDir, "classes/java/main")
+	enableArtemisPlugin = true
     enablePooledWeaving = true
     generateLinkMutators = true
     optimizeEntitySystems = true
 }
 
 compileJava.dependsOn fluid
-classes.finalizedBy weave
 
 // Help intellIJ pick up the generated classes.
 idea {
