+++ shared/src/shared/interfaces/Race.java
@@ -2,6 +2,10 @@ package shared.interfaces;
 
 import com.artemis.E;
 
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
 public enum Race {
     HUMAN,
     GNOME,
@@ -9,9 +13,11 @@ public enum Race {
     DROW,
     DWARF;
 
+    private static final List<Race> VALUES = Collections.unmodifiableList(Arrays.asList(values()));
+
     public static Race of(E entity) {
         int heroId = entity.getCharHero().heroId;
-        Hero hero = Hero.values()[heroId];
-        return values()[hero.getRaceId()];
+        Hero hero = Hero.VALUES.get(heroId);
+        return VALUES.get(hero.getRaceId());
     }
 }
